Update dynamic settings
Added default extra and store settings
